Update what kind of assertions are fired for bzip2
In passing, also update the bzip2 and gz default to be AssertError as
that's the more common case in both, and add a 'break;' to the default
clause- we don't intend to be just falling through those case
statements, even if the default is the last, we should be explicit about
that.

In passing, clean up some tabs that snuck in, rename a variable to be
more clear, and add some comments.