Add ASSERT_INLINE() macro.
When coverage testing ASSERT() macros in inline functions will be expanded and won't be recognized in our coverage rules that ignore ASSERT(). Since there are then uncovered conditions the coverage is incomplete.

The prior method required copying several lines of code and an explanatory comment into each inline function. Instead create a special macro for inclusion in inline functions.

Another possibility would be to automatically identify inline functions and add them to the coverage exclusions but that's an idea for another day.