Fix ill-advised cast in CipherBlock.
Little-endian architectures store the low-order bytes in the lowest memory location so this worked even in the case that size_t and int had different byte representations. Since buffer sizes are constrained there was no chance of the integer becoming negative and causing a problem that way.

On big-endian architectures this cast caused the low-order bytes to get loaded into the high-order bytes resulting in a huge buffer size that immediately triggered an assertion (and without the assertion would have certainly segfaulted).

Instead use a temporary int variable and cast that to size_t after the function call. This is the correct way to do it regardless of architecture.

This issue was detected while testing on the s390x architecture.