Prepare the configuration module for multi-repository support.
Refactor the code to allow a dynamic number of indexes for indexed options, e.g. pg-path. Our reliance on getopt_long() still limits the number of indexes we can have per group, but once this limitation is removed the rest of the code should be happy with dynamic numbers of indexes (with a reasonable maximum).

Allow the first key in a group to be any valid key, not just 1. Previously the first key, e.g. pg1-path had to be configured in any group. Now it is possible to configure, e.g. pg3-path, pg5-path, pg8-path. In the internal configuration structure there are no
longer gaps for missing keys, so the prior keys would map to index 0, 1, 2 internally.

Add an option to set a default in each group. This was previously handled by the host-id option but now there is a specific option for each group, pg-default and repo-default. These remain internal until they can be fully tested with multi-repo support. They are fully tested for internal usage.

Remove the ConfigDefineOption enum and use the ConfigOption enum instead. They are now equal since the indexed options (e.g. cfgOptRepoHost2) have been removed from ConfigOption.

Remove the config/config test module and add required tests to the config/parse test module. Parsing is now the only way to load a config so this removes some redundancy.

Split new internal config structures and functions into a new header file, config.intern.h. More functions will need to be moved over from config.h but that will need to be done in a future commit to reduce churn.